kafka/app/requirements.txt
certifi==2019.9.11
chardet==3.0.4
future==0.18.2
idna==2.8
kafka-python==1.4.6
oauthlib==3.1.0
PySocks==1.7.1
python-twitter==3.5
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
tweepy==3.8.0
urllib3==1.25.7
